:rocket: feat: annotate tests with @Tag and restructure packages
