+++ src/test/java/dev/journey/restcountries/unit/controller/CountryControllerTest.java
@@ -1,9 +1,11 @@
-package dev.journey.restcountries.controller;
+package dev.journey.restcountries.unit.controller;
 
+import dev.journey.restcountries.controller.CountryController;
 import dev.journey.restcountries.dto.CountryResponse;
-import dev.journey.restcountries.factory.CountryInstancioFactory;
+import dev.journey.restcountries.unit.factory.CountryInstancioFactory;
 import dev.journey.restcountries.service.CountryServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +15,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+@Tag("unit")
 class CountryControllerTest {
 
     private CountryServiceImpl countryService;
